add ut for release-7.1-keyspace (#539)
* add ut


* comment some ut


* comment some ut


* comment some ut


* bazel_coverage_test


* make bazel_prepare


* fix make bazel_coverage_test


* comment some ut


* fix main_test


* revert ut


* revert ut


* make check


* make check


* make test


* make test


* add make test part1


* add make test part1


* add make test part1


* add make test part1


* add make part2


* add more test part


* use cache


* make ut


* skip br/dumpling/parser


* skip br/dumpling/parser


* make ut better


* remote fast-fail


* make ut better


* make unit test better


* make unit test better


* make ut better


* make check


* make check


* make ut better


* make ut better


* make ut better


* make ut better


* make ut better


* make ut better


* make ut better


* make ut better


* make ut better


* make check


* make check


* make ut better


* make ut better


* use cache


* use cache


* make ut better


* make ut better


* make ut better


* make ut better


* make ut better


* use intest.test to determine if running in test


* executor pkg test compatible with go1.21


* skip inconsistent variance calculation


* fix analyze test:


* fix executor test


* fmt


* fix TestTiDBClusterInfo


* fix TestGlobalMemoryControl


* fix TestTiDBVersion


* modify TestVariablesInfo for serverless


* fix TestMemoryUsageAndOpsHistory


* fix previleges_test


* make check


* fix conn_stmt_test.go


* fix TestGracefulShutdown


* fix TestTiDBServerMemoryLimit2


* fix update_test.go


* fix sem_test.go


* fix reader_test.go


* enable mpp test


* fix codec_test.go


* fix GetInstance


* improve stale read test stability


* make check


* remove test code


* specify TZ in makefile


* re-enable ci test


* make check


* force single thread test execution


* try make ut


* retry matrix


* change worker count to 4


* try use bazel


* keep test going in the event of failure


* use make ut


* force single thread test execution in 16 parts


* don't clean test binaries


* fix typo


* skip checking for mpp task generation


* skip second check of lock aquire


* improve TestCursorFetchErrorInFetch stability


* cleanup


* cleanup


* optimize matrix, add concurrency restriction


* fix typo


* use max-parallel


* test concurrency group cancel


* retry when connection refused


* increase concurrency


* optimize test build


* increase concurrency


* split ut so it only trigger when commented


* manual trigger


* remove manual trigger


* trigger on pr for now


* add concurrency and retry


* split large package into separate part


* move executor to its separate part


* re enable TestGlobalMemoryControl


* make check


* address comments


* fmt


* time zone has to be in ci


* improve config test stability


* remove empty line


* make test get lock more stable


---------

+++ br/pkg/lightning/config/config.go
@@ -991,9 +991,9 @@ func NewConfig() *Config {
 		},
 		RUConfig: RUConfig{
 			ReportWRU:             true,
-			WriteBaseCost:         float64(defaultRUConfig.WriteBaseCost),
-			WritePerBatchBaseCost: float64(defaultRUConfig.WritePerBatchBaseCost),
-			WriteCostPerByte:      float64(defaultRUConfig.WriteCostPerByte),
+			WriteBaseCost:         defaultRUConfig.WriteBaseCost,
+			WritePerBatchBaseCost: defaultRUConfig.WritePerBatchBaseCost,
+			WriteCostPerByte:      defaultRUConfig.WriteCostPerByte,
 		},
 		Metrics: Metrics{
 			Addrs:    nil,
